Projects/Project3/requirements.txt
img2vec-pytorch
torchvision
pillow
numpy
scikit-learn
opencv-python
torch
optuna
xgboost
matplotlib